gdrive/requirement.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
distlib==0.3.6
filelock==3.9.0
google-api-core==2.11.0
google-api-python-client==2.73.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
numpy==1.24.1
oauth2client==4.1.3
oauthlib==3.2.2
pandas==1.5.3
platformdirs==2.6.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.17.1
